Board And Cell Classes
Added a Board and Cell class.
The Cell class holds a value and x and y position.
The Board class holds an array of 7x6 of cells, and can currently make moves. There is an unfinished function to fill the form with data from the board, but we need to set up the form to get that working.
Also added a board into the Form1.cs, anyone feel free to get rid of that if we want it on a different form. Most of my changes on there are just for showing what I'm thinking with the game.